# openapi.json
{"openapi":"3.0.2","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/v1/register/email":{"post":{"summary":"Register Via Email","operationId":"register_via_email_v1_register_email_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassicUserPost"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/login/email":{"post":{"summary":"Login Via Email","operationId":"login_via_email_v1_login_email_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClassicLoginUser"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/user":{"get":{"summary":"Get Users","operationId":"get_users_v1_user_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/user/edit":{"post":{"summary":"Update Users","operationId":"update_users_v1_user_edit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModifyUserPost"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/testapi":{"get":{"summary":"Test","operationId":"test_v1_testapi_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/validate/token":{"get":{"summary":"Validate Token","operationId":"validate_token_v1_validate_token_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"JWTBearer":[]}]}},"/v1/user/current":{"get":{"summary":"Get Current User","operationId":"get_current_user_v1_user_current_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"JWTBearer":[]}]}},"/v1/":{"get":{"summary":"Index","operationId":"index_v1__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/info":{"get":{"summary":"Info","operationId":"info_v1_info_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/lessons/search":{"get":{"summary":"Search Lesson","operationId":"search_lesson_v1_lessons_search_get","parameters":[{"required":true,"schema":{"title":"Term","type":"string"},"name":"term","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/lessons":{"get":{"summary":"Get Lessons","operationId":"get_lessons_v1_lessons_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/lessons{lesson_id}":{"get":{"summary":"Get Lesson By Id","operationId":"get_lesson_by_id_v1_lessons_lesson_id__get","parameters":[{"required":true,"schema":{"title":"Lesson Id","type":"string"},"name":"lesson_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/lessons/create":{"post":{"summary":"Post Lessons","operationId":"post_lessons_v1_lessons_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LessonPost"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/lessons/grade":{"get":{"summary":"Get Grade By User Id Lesson Id","operationId":"get_grade_by_user_id_lesson_id_v1_lessons_grade_get","parameters":[{"required":true,"schema":{"title":"User Id","type":"string"},"name":"user_id","in":"query"},{"required":true,"schema":{"title":"Lesson Id","type":"string"},"name":"lesson_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Post Grade","operationId":"post_grade_v1_lessons_grade_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GradeDTO"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ClassicLoginUser":{"title":"ClassicLoginUser","required":["email","password"],"type":"object","properties":{"email":{"title":"Email","type":"string"},"password":{"title":"Password","type":"string"}}},"ClassicUserPost":{"title":"ClassicUserPost","required":["name","password","email"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"password":{"title":"Password","type":"string"},"email":{"title":"Email","type":"string"}}},"GradeDTO":{"title":"GradeDTO","required":["user_id","lesson_id","grade","completed"],"type":"object","properties":{"user_id":{"title":"User Id","type":"string"},"lesson_id":{"title":"Lesson Id","type":"string"},"grade":{"title":"Grade","type":"number"},"completed":{"title":"Completed","type":"boolean"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"LessonPost":{"title":"LessonPost","required":["title","description","example_video","category_name"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"example_video":{"title":"Example Video","type":"string"},"category_name":{"title":"Category Name","type":"string"}}},"ModifyUserPost":{"title":"ModifyUserPost","required":["id","name","email","score"],"type":"object","properties":{"id":{"title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"email":{"title":"Email","type":"string"},"score":{"title":"Score","type":"number"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"JWTBearer":{"type":"http","scheme":"bearer"}}}}